fix(ci): a scan root that cannot be enumerated is a refusal, not a clean result
`findFiles` shelled out through `execSync` with Node's default 1 MiB
`maxBuffer`. A root whose `find` output exceeds that raises ENOBUFS, which
landed in a catch that logged and continued, so the root contributed ZERO
files and the run still printed a clean result.

Measured, with the buffer forced to 512 bytes so the condition is real:

  before   Found 8 files to process
           All files have the license header.        exit 0
  after    Could not enumerate scan roots:
             apps: spawnSync /bin/sh ENOBUFS
             examples: spawnSync /bin/sh ENOBUFS     exit 2

Eight files of 5,378, announced as full coverage. A gate that under-scans and
reports success is the exact failure this one exists to prevent, so it shipped
with one inside it.

Fixed both ways, because either alone is not enough: an explicit 64 MiB buffer
so ordinary growth cannot trip it, AND enumeration failures recorded and made
fatal with exit 2, matching how a missing root is already treated. Raising the
buffer alone only moves the cliff.

Enumeration moved to `scripts/lib/license-scan.mjs`. The fix took the CLI to
432 lines, over the 400-line house rule, and the seam is a real one: deciding
WHICH files a run covers is a different job from deciding what to do with them,
and this half's failures are all "the run covered less than it claims".

From CodeRabbit on #4217.
